# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Firetower",
    "description": "The Firetower control plane: API, scheduling, and worker transports.",
    "contact": {
      "name": "Westlabs LLC"
    },
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.10.0"
  },
  "paths": {
...
    "/api/v1/secrets/{scope}/{name}": {
      "put": {
        "tags": [
          "secrets"
        ],
        "summary": "Replace a credential with a new one.",
        "description": "Only for a name that already exists. Storing under an arbitrary name would\nlet this screen fill up with values nothing ever reads — what a credential is\n*for* is decided where it is used, not here.",
        "operationId": "replace_secret",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "description": "Secret scope",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Secret name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceSecret"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "secrets"
        ],
        "summary": "Remove a credential.",
        "description": "Whatever used it goes back to having none — an agent shows as needing one\nagain, a git host shows as not authorized. The log entry stays.",
        "operationId": "remove_secret",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "description": "Secret scope",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Secret name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/secrets/{scope}/{name}/reveal": {
      "post": {
        "tags": [
          "secrets"
        ],
        "summary": "Put a credential on screen.",
        "description": "A `POST` because it changes something: it writes a `Reveal` into the access\nlog, which is what makes this defensible at all. A `GET` would also sit in\nbrowser history and proxy logs, which a credential shouldn't.\n\nThis is the one route that hands a stored value back, and it exists because\na credential you cannot inspect is one you cannot verify or copy elsewhere.\nThe cost is real: anything that can reach this API can read every token\nhere, and the log is what is left to notice it.",
        "operationId": "reveal_secret",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "description": "Secret scope",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Secret name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevealedSecret"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sessions/{id}/pty": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "The terminal.",
        "description": "A websocket rather than the event stream: this is the one thing in Firetower\nthat genuinely flows both ways, byte at a time, and where latency is felt.\n\nOutput arrives as binary frames of raw terminal bytes. Input goes back the\nsame way; a text frame is a control message, which today means resizing.",
        "operationId": "session_pty",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cols",
            "in": "query",
            "description": "Terminal width",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "rows",
            "in": "query",
            "description": "Terminal height",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "shell",
            "in": "query",
            "description": "A shell of your own rather than the agent's terminal",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Terminal stream"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Every non-success response, so failures are as typed as everything else.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "description": "For humans and logs. The interface should switch on `code` and write its\nown copy — only it knows the context and what to offer next."
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "The catalogue is the type, so there is no separate list to keep in sync.",
        "enum": [
          "InvalidRequest",
          "NotFound",
          "NoCapacity",
          "HostUnreachable",
          "RepoNotConnected",
          "SessionEnded",
          "ProviderNotConfigured",
          "ProviderNotConnected",
          "RepoAccessDenied",
          "RepoUnreachable",
          "RepoUnusable",
          "RepoInUse",
          "ActionFailed",
          "Unauthorized",
          "PasswordChangeRequired",
          "Internal"
        ]
      },
...
      "ReplaceSecret": {
        "type": "object",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "type": "string",
            "description": "What to store from now on. The previous value is not recoverable."
          }
        }
      },
...
      "RevealedSecret": {
        "type": "object",
        "description": "A value, on its way to a person who asked for it.",
        "required": [
          "scope",
          "name",
          "value"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
    }
  }
}